fix: invalid casts and conversion
narrowing conversions are dangerous and this commit aims to remove a few
of them
many can be seen when converting unsigned longs to ints
this narrows down the width from 64bit to 32bit (losing half of the
bits)

+++ src/functions.c
@@ -1,15 +1,15 @@
 #include "header.h"
 
-void stotime(int total_seconds, int *hours, int *minutes, int *seconds)
+void stotime(unsigned long total_seconds, uint8_t *hours, uint8_t *minutes, uint8_t *seconds)
 {
-	int seconds_remainder = total_seconds % 60;
+	unsigned long seconds_remainder = total_seconds % 60;
 	*seconds = seconds_remainder;
 
-	int minutes_pre_remainder = total_seconds - seconds_remainder;
-	int minutes_remainder = minutes_pre_remainder % 3600;
+	unsigned long minutes_pre_remainder = total_seconds - seconds_remainder;
+	unsigned long minutes_remainder = minutes_pre_remainder % 3600;
 	*minutes = minutes_remainder / 60;
 
-	int hours_temp = minutes_pre_remainder - minutes_remainder;
+	unsigned long hours_temp = minutes_pre_remainder - minutes_remainder;
 	*hours = hours_temp / 3600;
 }
